% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alteracoes_rcd202531}
\alias{alteracoes_rcd202531}
\title{Collection Structure Changes (RCD 2025/31)}
\format{
A tibble with 31 rows and 7 columns:
\describe{
\item{municipio_nome}{Municipality name}
\item{agencia_nome_original}{Original agency name}
\item{agencia_nome_nova}{New agency name}
\item{alteracao_diarias}{Change in daily allowance requirement}
\item{municipio_codigo}{Municipality code}
\item{agencia_codigo_original}{Original agency code}
\item{agencia_codigo_nova}{New agency code}
}
}
\usage{
alteracoes_rcd202531
}
\description{
Records of municipality reassignments between IBGE agencies following
administrative change RCD 2025/31.
}
\keyword{datasets}
