% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assistencias_ba}
\alias{assistencias_ba}
\title{Agency Assistance Areas (Bahia)}
\format{
A tibble with 50 rows and 2 columns:
\describe{
\item{agencia_codigo}{Agency code}
\item{assistencia_nome}{Assistance area name}
}
}
\usage{
assistencias_ba
}
\description{
Maps IBGE agencies in Bahia to their assigned assistance/support area names.
}
\keyword{datasets}
