% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distancias_agencias_municipios_osrm}
\alias{distancias_agencias_municipios_osrm}
\title{Agency–Municipality Road Distances (OSRM)}
\format{
A data frame with 222,187 rows and 4 columns:
\describe{
\item{agencia_codigo}{Agency code}
\item{municipio_codigo}{Municipality code}
\item{distancia_km}{Road distance in kilometers}
\item{duracao_horas}{Travel duration in hours}
}
}
\source{
Open Source Routing Machine (OSRM)
}
\usage{
distancias_agencias_municipios_osrm
}
\description{
Road distances and travel durations between all IBGE agencies and
Brazilian municipalities, calculated via OSRM routing.
}
\keyword{datasets}
