% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipios}
\alias{municipios}
\title{Brazilian Municipalities (Spatial)}
\format{
A \code{sf} data frame with columns:
\describe{
\item{municipio_codigo}{Municipality code (7 digits)}
\item{municipio_sede_lon}{Longitude of the municipal seat}
\item{municipio_sede_lat}{Latitude of the municipal seat}
\item{municipio_nome}{Municipality name}
\item{uf_codigo}{State code}
\item{uf_sigla}{State abbreviation}
\item{uf_nome}{State name}
\item{regiao_codigo}{Region code}
\item{municipio_populacao}{Estimated population (SIDRA table 6579)}
}
}
\source{
IBGE geographic base (geobr) and SIDRA population estimates
}
\usage{
municipios
}
\description{
Spatial data with municipal seat locations, names, and population
estimates for all Brazilian municipalities.
}
\keyword{datasets}
