% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipios_22}
\alias{municipios_22}
\title{Brazilian Municipalities 2022 (Spatial)}
\format{
A \code{sf} data frame with 5,571 rows and 11 columns:
\describe{
\item{municipio_codigo}{Municipality code (7 digits)}
\item{municipio_sede_lon}{Longitude of the municipal seat}
\item{municipio_sede_lat}{Latitude of the municipal seat}
\item{municipio_nome}{Municipality name}
\item{uf_codigo}{State code}
\item{uf_sigla}{State abbreviation}
\item{uf_nome}{State name}
\item{regiao_codigo}{Region code}
}
}
\source{
IBGE geographic base 2022
}
\usage{
municipios_22
}
\description{
Spatial boundaries and metadata for Brazilian municipalities from the
2022 IBGE geographic base.
}
\keyword{datasets}
