% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipios_codigos}
\alias{municipios_codigos}
\title{Municipality Administrative Codes}
\format{
A tibble with 5,570 rows and 4 columns:
\describe{
\item{municipio_codigo}{Municipality code (7 digits)}
\item{micro_codigo}{Microregion code}
\item{aglomeracao_codigo}{Urban agglomeration code}
\item{data}{Reference date}
}
}
\source{
IBGE geographic base 2022
}
\usage{
municipios_codigos
}
\description{
Reference table linking municipality codes to microregion and urban
agglomeration codes, with reference date.
}
\keyword{datasets}
