% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrm_local.R
\name{osrm_local_clear_cache}
\alias{osrm_local_clear_cache}
\title{Clear Cached PBF Files}
\usage{
osrm_local_clear_cache()
}
\value{
A character vector of deleted file paths (invisibly).
}
\description{
Removes downloaded \code{.osm.pbf} files from the orcedata cache directory.
}
