% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrm_local.R
\name{osrm_local_status}
\alias{osrm_local_status}
\title{Check Local OSRM Server Status}
\usage{
osrm_local_status()
}
\value{
\code{TRUE} if a server is running, \code{FALSE} otherwise (invisibly).
}
\description{
Reports whether a local OSRM server is currently running.
}
