% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrm_local.R
\name{osrm_local_stop}
\alias{osrm_local_stop}
\title{Stop the Local OSRM Server}
\usage{
osrm_local_stop()
}
\value{
\code{NULL} (invisibly).
}
\description{
Stops all running OSRM servers and restores the previous \code{osrm.server}
option.
}
