% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ufs}
\alias{ufs}
\title{Brazilian States (UFs)}
\format{
A \code{sf} data frame with 27 rows and 8 columns:
\describe{
\item{uf_codigo}{State code}
\item{uf_sigla}{State abbreviation}
\item{uf_nome}{State name}
\item{regiao_codigo}{Region code}
\item{regiao_nome}{Region name}
\item{uf_lon}{Longitude of the state representative point}
\item{uf_lat}{Latitude of the state representative point}
\item{geom}{Polygon geometry (sf)}
}
}
\source{
IBGE geographic base 2022
}
\usage{
ufs
}
\description{
Spatial boundaries and metadata for Brazilian states (Unidades Federativas).
}
\keyword{datasets}
